Split advisory client protocol boundaries
