+++ quality/ci_quality_gates.md
@@ -62,12 +62,12 @@ Most recent local PR-grade evidence:
 
 1. The previous quality-baseline enforcement branch added
    `scripts/check_refactor_quality_thresholds.py` as a blocking lint-stage gate.
-2. Current enforced source-file threshold: no Python source file under `src/app` above 630
+2. Current enforced source-file threshold: no Python source file under `src/app` above 628
    script-counted lines.
 3. Current enforced function threshold: no Python function or async function above the remediated
    49-line AST span baseline.
 4. `python scripts/check_refactor_quality_thresholds.py`: passed with
-   `max_source_file_lines=630` and `max_function_lines=49`.
+   `max_source_file_lines=628` and `max_function_lines=49`.
 5. Feature Lane and PR Merge Gate step names now call out `Lint and Refactor Quality Thresholds`
    so the promoted gate is visible in GitHub logs.
 6. Current advisor-brief contract-boundary branch moves Advisor Brief presentation/source item
@@ -115,6 +115,28 @@ Most recent local PR-grade evidence:
    records 13/13 DNS checks, 13/13 representative API checks, 4/4 metric checks, 14 log artifacts,
    and 5/5 observability screenshots. GitHub checks and post-merge wiki publication remain pending
    for this branch.
+9. Current advisory protocol-boundary branch splits the mixed advisory client protocol surface into
+   focused bank-demo proof, copilot, workspace, cockpit, policy, and proposal protocol modules
+   while preserving the public `app.services.advisory_client_protocols` facade. Focused validation
+   passed with ruff check, mypy over `src`, 67 protocol/service-boundary tests, and
+   refactor-threshold trials proving `max_source_file_lines=628` passes while `627` fails only on
+   `src/app/clients/dpm_wave_client.py`. Full local `make check` passed with ruff, format check
+   over 779 files, monetary-float guard, refactor threshold gate, workflow action-runtime baseline,
+   mypy over 556 source files, OpenAPI smoke, and 1,239 unit/contract tests. Full local `make ci`
+   passed with migration contract smoke, 209 integration tests, 1,448 combined coverage tests,
+   94.30% total coverage, and no known vulnerabilities after the governed `PYSEC-2026-161`
+   exception. Branch-specific live canonical validation passed after rebuilding the Docker-backed
+   Gateway and downstream stack, then rerunning after performance lineage materialization completed:
+   `lotus-workbench/output/playwright/live-canonical-advisory-protocol-boundaries-rerun/live-validation-summary.json`
+   records 95 API checks, 2 calculation checks, 29 screenshots, 25/25 ready panel classifications,
+   28 supportability checks, 10 workflow-pack checks, no missing or non-ready panels, 9/9 RFC36-43
+   features validated, and 0 RFC36-43 gaps. Companion observability evidence at
+   `lotus-workbench/output/observability-live/advisory-protocol-boundaries-rerun/observability-evidence-manifest.json`
+   records 13/13 DNS checks, 13/13 representative API checks, 4/4 metric checks, 14 log artifacts,
+   5/5 observability screenshots, Gateway correlation/request/trace IDs, 58 fan-out events, and 14
+   audit events. Residual data-mesh limitation is not Gateway-owned: performance contribution source
+   economics remains `SOURCE_LIMITED` for non-source-authored component P&L economics. GitHub checks
+   and post-merge wiki publication remain pending for this branch.
 7. Current DPM command-center exception-summary boundary branch focused validation passed with
    ruff check, ruff format check, mypy over four touched DPM command-center service modules,
    74 focused DPM command-center service/router/boundary tests, and refactor-threshold trials
