+++ quality/refactor_health_report.md
@@ -646,19 +646,25 @@ cache-status/correlation stamping into `src/app/services/risk_workspace_cache.py
 `risk_workspace_service.py` below the source-file ceiling. The blocking source-file threshold is
 ratcheted from 632 to 630 lines, with `src/app/services/advisory_client_protocols.py` now the
 single source-file ceiling blocker.
+The current advisory protocol-boundary slice splits bank-demo proof, advisory copilot, advisory
+workspace, advisor cockpit, advisory policy, and proposal protocol families into focused modules
+while preserving the public `app.services.advisory_client_protocols` compatibility facade. Advisory
+services now import focused protocol families directly. The blocking source-file threshold is
+ratcheted from 630 to 628 lines, with `src/app/clients/dpm_wave_client.py` now the single
+source-file ceiling blocker.
 
 ## Health Signals
 
 | Area | Current posture | Evidence |
 | --- | --- | --- |
-| Branch hygiene | Healthy | Current risk workspace cache-boundary branch was created from clean `main`; no open PRs and no unmerged remote branches were present at slice start; final remote/local cleanup remains a post-merge gate |
-| Unit/contract coverage | Healthy | Focused risk workspace cache-boundary validation passed with 54 risk workspace cache/service/boundary/threshold tests; full local `make check` passed with 1,236 unit/contract tests, and refactor-threshold trials prove `max_source_file_lines=630` passes while `629` fails only on `src/app/services/advisory_client_protocols.py` |
+| Branch hygiene | Healthy | Current advisory protocol-boundary branch was created from clean `main`; stranded-truth reconciliation found no unmerged remote branches; final remote/local cleanup remains a post-merge gate |
+| Unit/contract coverage | Healthy | Focused advisory protocol-boundary validation passed with 67 protocol/service-boundary tests, ruff, mypy over `src`, and refactor-threshold trials proving `max_source_file_lines=628` passes while `627` fails only on `src/app/clients/dpm_wave_client.py`; full local `make check` passed with 1,239 unit/contract tests |
 | Integration coverage | Healthy | Full local `make ci` passed with migration contract smoke and 209 integration tests |
-| Total coverage | Healthy | Full local `make ci` passed with 1,445 combined coverage tests and 94.33% total coverage |
-| Security audit | Governed | Current risk workspace cache-boundary branch introduces no dependency, authentication, caller-context, product-error-detail, upstream error-shape, monetary-float conversion, or data-mesh behavior changes; full local `make ci` passed `pip-audit` with no known vulnerabilities after the governed `PYSEC-2026-161` exception |
-| Modularity | Improving, incomplete | Longest-function baseline remains 49 lines; current branch extracts risk workspace cache policy into `src/app/services/risk_workspace_cache.py`, reducing `risk_workspace_service.py` below the ceiling. The largest current source-file hotspot is now `src/app/services/advisory_client_protocols.py` at 630 script-counted lines |
-| Live canonical runtime | Healthy for this branch | Branch-specific canonical proof passed after rebuilding the Docker-backed Gateway and downstream stack. Machine-readable evidence at `lotus-workbench/output/playwright/live-canonical-risk-workspace-cache-boundary/live-validation-summary.json` records 94 API checks, 2 calculation checks, 29 screenshots, 25/25 ready panel classifications, 28 supportability checks, 10 workflow-pack checks, no missing or non-ready panels, 9/9 RFC36-43 features validated, and 0 RFC36-43 gaps |
-| Observability evidence | Healthy for this branch | Branch-specific companion pack `lotus-workbench/output/observability-live/risk-workspace-cache-boundary-20260619/observability-evidence-manifest.json` captured canonical DNS, representative API samples, metrics, bounded logs, and observability screenshots; 13/13 DNS checks passed, 13/13 representative API checks returned HTTP 200, 4/4 metric checks returned HTTP 200, 14 log artifacts were captured, and 5/5 observability screenshots returned HTTP 200 |
+| Total coverage | Healthy | Full local `make ci` passed with 1,448 combined coverage tests and 94.30% total coverage |
+| Security audit | Governed | Current advisory protocol-boundary branch introduces no dependency, authentication, caller-context, product-error-detail, upstream error-shape, monetary-float conversion, data-mesh behavior, or runtime behavior changes; full local `make ci` passed `pip-audit` with no known vulnerabilities after the governed `PYSEC-2026-161` exception |
+| Modularity | Improving, incomplete | Longest-function baseline remains 49 lines; current branch splits advisory client protocol families into focused modules and preserves a compatibility facade. The largest current source-file hotspot is now `src/app/clients/dpm_wave_client.py` at 628 script-counted lines |
+| Live canonical runtime | Healthy for this branch | Branch-specific canonical proof passed after rebuilding the Docker-backed Gateway and downstream stack, then rerunning validation after performance lineage materialization completed. Machine-readable evidence at `lotus-workbench/output/playwright/live-canonical-advisory-protocol-boundaries-rerun/live-validation-summary.json` records 95 API checks, 2 calculation checks, 29 screenshots, 25/25 ready panel classifications, 28 supportability checks, 10 workflow-pack checks, no missing or non-ready panels, 9/9 RFC36-43 features validated, and 0 RFC36-43 gaps |
+| Observability evidence | Healthy with residual data-mesh qualification | Branch-specific companion pack `lotus-workbench/output/observability-live/advisory-protocol-boundaries-rerun/observability-evidence-manifest.json` captured canonical DNS, representative API samples, metrics, bounded logs, and observability screenshots; 13/13 DNS checks passed, 13/13 representative API checks returned HTTP 200, 4/4 metric checks returned HTTP 200, 14 log artifacts were captured, and 5/5 observability screenshots returned HTTP 200. Gateway log review found correlation/request/trace identifiers, 58 fan-out events, 14 audit events, and no ERROR/5xx scan hits. Remaining data-mesh qualification is upstream/domain-owned: performance contribution source economics reports `SOURCE_LIMITED` for non-source-authored component P&L economics |
 | API governance | Improving, incomplete | 233 OpenAPI paths and 247 operations have summaries, descriptions, operation IDs, tags, and documented 4xx/5xx responses; Spectral remains report-only |
 | Error consistency | Improving, incomplete | Reporting job and report-batch upstream error handling now uses explicit code-owned mapping rules with focused product-safe fallback tests; shared generic service-error status mapping is code-owned and tested; advisory-facing product-safe service-error defaults now use typed immutable configs; broader route/upstream error normalization remains open |
 | Architecture rules | Improving, incomplete | AST boundary tests exist; import-linter is report-only; source-file and function-size thresholds are now blocking through `make lint` |
