+++ src/app/services/advisory_client_protocols.py
@@ -1,630 +1,15 @@
-from typing import Any, Protocol
-
-
-class BankDemoProofClient(Protocol):
-    async def get_bank_demo_proof_scenario_contract(
-        self,
-        *,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_bank_demo_supported_claim_register(
-        self,
-        *,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def build_bank_demo_proof_pack(
-        self,
-        *,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-
-class AdvisoryCopilotClient(Protocol):
-    async def create_advisory_copilot_evidence_packet(
-        self,
-        *,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_advisory_copilot_evidence_packet_from_proposal_version(
-        self,
-        *,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_advisory_copilot_evidence_packet(
-        self,
-        *,
-        evidence_packet_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def run_advisory_copilot_action(
-        self,
-        *,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_advisory_copilot_run(
-        self,
-        *,
-        run_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def review_advisory_copilot_run(
-        self,
-        *,
-        run_id: str,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_advisory_copilot_supportability(
-        self,
-        *,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_advisory_copilot_proposal_version_runs(
-        self,
-        *,
-        proposal_id: str,
-        version_id: str,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-
-class AdvisoryWorkspaceClient(Protocol):
-    async def create_advisory_workspace(
-        self,
-        *,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_advisory_workspace(
-        self,
-        *,
-        workspace_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def apply_advisory_workspace_draft_action(
-        self,
-        *,
-        workspace_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def evaluate_advisory_workspace(
-        self,
-        *,
-        workspace_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def save_advisory_workspace(
-        self,
-        *,
-        workspace_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_advisory_workspace_saved_versions(
-        self,
-        *,
-        workspace_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_advisory_workspace_saved_version_replay_evidence(
-        self,
-        *,
-        workspace_id: str,
-        workspace_version_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def resume_advisory_workspace(
-        self,
-        *,
-        workspace_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def compare_advisory_workspace(
-        self,
-        *,
-        workspace_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def request_advisory_workspace_rationale(
-        self,
-        *,
-        workspace_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def review_advisory_workspace_rationale(
-        self,
-        *,
-        workspace_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def handoff_advisory_workspace(
-        self,
-        *,
-        workspace_id: str,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-
-class AdvisorCockpitClient(Protocol):
-    async def list_advisor_cockpit_actions(
-        self,
-        *,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_advisor_cockpit_preparation_packets(
-        self,
-        *,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_advisor_cockpit_action(
-        self,
-        *,
-        action_item_id: str,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_advisor_cockpit_snapshot(
-        self,
-        *,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_advisor_cockpit_supportability(
-        self,
-        *,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def acknowledge_advisor_cockpit_action(
-        self,
-        *,
-        action_item_id: str,
-        body: dict[str, Any],
-        params: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def evaluate_advisor_cockpit_house_view_cohort(
-        self,
-        *,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-
-class AdvisoryPolicyClient(Protocol):
-    async def list_policy_packs(
-        self,
-        *,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_policy_pack_version(
-        self,
-        *,
-        policy_pack_id: str,
-        policy_version: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def validate_policy_pack_version(
-        self,
-        *,
-        policy_pack_id: str,
-        policy_version: str,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def activate_policy_pack_version(
-        self,
-        *,
-        policy_pack_id: str,
-        policy_version: str,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_policy_evaluation(
-        self,
-        *,
-        proposal_id: str,
-        proposal_version_id: str,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_policy_review_queue(
-        self,
-        *,
-        evaluation_status: str | None,
-        portfolio_id: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_policy_evaluation(
-        self,
-        *,
-        evaluation_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def replay_policy_evaluation(
-        self,
-        *,
-        evaluation_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def record_policy_evaluation_event(
-        self,
-        *,
-        evaluation_id: str,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_policy_evaluation_lineage(
-        self,
-        *,
-        evaluation_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_policy_sign_off_package(
-        self,
-        *,
-        evaluation_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_policy_evaluation_workflow(
-        self,
-        *,
-        evaluation_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def record_policy_sign_off_decision(
-        self,
-        *,
-        evaluation_id: str,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def request_policy_report_package(
-        self,
-        *,
-        evaluation_id: str,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def request_policy_ai_evidence(
-        self,
-        *,
-        evaluation_id: str,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-
-class ProposalClient(Protocol):
-    async def simulate_proposal(
-        self,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_proposal_artifact(
-        self,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_proposal(
-        self,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_proposals(
-        self,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal(
-        self,
-        proposal_id: str,
-        include_evidence: bool,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_version(
-        self,
-        proposal_id: str,
-        version_no: int,
-        include_evidence: bool,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_proposal_version(
-        self,
-        proposal_id: str,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_proposal_async(
-        self,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_proposal_version_async(
-        self,
-        proposal_id: str,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_operation(
-        self,
-        operation_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_operation_by_correlation(
-        self,
-        operation_correlation_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_operation_replay_evidence(
-        self,
-        operation_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def transition_proposal(
-        self,
-        proposal_id: str,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def record_approval(
-        self,
-        proposal_id: str,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_workflow_events(
-        self,
-        proposal_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_approvals(
-        self,
-        proposal_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_lineage(
-        self,
-        proposal_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_version_replay_evidence(
-        self,
-        proposal_id: str,
-        version_no: int,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_idempotency_record(
-        self,
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def regenerate_proposal_narrative(
-        self,
-        proposal_id: str,
-        version_no: int,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_narrative(
-        self,
-        proposal_id: str,
-        version_no: int,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def review_proposal_narrative(
-        self,
-        proposal_id: str,
-        version_no: int,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_report_request(
-        self,
-        proposal_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_execution_handoff(
-        self,
-        proposal_id: str,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_delivery_summary(
-        self,
-        proposal_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_delivery_events(
-        self,
-        proposal_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_execution_status(
-        self,
-        proposal_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def record_execution_update(
-        self,
-        proposal_id: str,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_proposal_memo(
-        self,
-        proposal_id: str,
-        version_no: int,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_memo(
-        self,
-        proposal_id: str,
-        version_no: int,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_memo_projection(
-        self,
-        proposal_id: str,
-        version_no: int,
-        audience: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def review_proposal_memo(
-        self,
-        proposal_id: str,
-        version_no: int,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def record_proposal_memo_report_package_event(
-        self,
-        proposal_id: str,
-        version_no: int,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def request_proposal_memo_report_package(
-        self,
-        proposal_id: str,
-        version_no: int,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def request_proposal_memo_ai_commentary(
-        self,
-        proposal_id: str,
-        version_no: int,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_memo_lineage(
-        self,
-        proposal_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_memo_replay_evidence(
-        self,
-        proposal_id: str,
-        version_no: int,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
+from app.services.advisor_cockpit_client_protocols import AdvisorCockpitClient
+from app.services.advisory_copilot_client_protocols import AdvisoryCopilotClient
+from app.services.advisory_policy_client_protocols import AdvisoryPolicyClient
+from app.services.advisory_workspace_client_protocols import AdvisoryWorkspaceClient
+from app.services.bank_demo_proof_client_protocols import BankDemoProofClient
+from app.services.proposal_client_protocols import ProposalClient
+
+__all__ = [
+    "AdvisorCockpitClient",
+    "AdvisoryCopilotClient",
+    "AdvisoryPolicyClient",
+    "AdvisoryWorkspaceClient",
+    "BankDemoProofClient",
+    "ProposalClient",
+]
