+++ src/app/services/bank_demo_proof_service.py
@@ -1,7 +1,7 @@
 from typing import Any
 
 from app.contracts.bank_demo_proof import BankDemoProofEnvelopeResponse
-from app.services.advisory_client_protocols import BankDemoProofClient
+from app.services.bank_demo_proof_client_protocols import BankDemoProofClient
 from app.services.upstream_envelope import (
     ProductSafeServiceErrorConfig,
     build_gateway_envelope,
